Revert "Standadizing suffixes to .hh and .cc"
This reverts commit e18b94731249e603c94ac25da49e0b7c58f0729b.
